abstract utility_cog and senders, refactor cog to welcome_and_coc_cog
