+++ README.md
@@ -75,11 +75,14 @@ docker-compose up -d
 
 - `bot/`: Main bot code
   - `__main__.py`: Entry point
-  - `cog.py`: Core bot functionality
-  - `models.py`: Database models
-  - `db.py`: Database connection management
   - `config.py`: Configuration handling
+  - `db.py`: Database connection management
   - `exceptions.py`: Custom exceptions
+  - `messages.py`: Messages sent to members based on actions
+  - `models.py`: Database models
+  - `senders.py`: Sends dms, creates private categories and channels if dms are closed
+  - `utility_cog.py`: Administration commands
+  - `welcome_and_coc_cog.py`: Actions related to new members joining
 - `tests/`: Test suite
 - `alembic/`: Database migrations
