+++ bot/__main__.py
@@ -3,7 +3,7 @@ import asyncio
 import discord
 from discord.ext import commands
 
-from bot import cog, config
+from bot import welcome_and_coc_cog, utility_cog, config
 
 
 async def main() -> None:
@@ -14,7 +14,8 @@ async def main() -> None:
 
     discord.utils.setup_logging()
     bot = commands.Bot(command_prefix="!", intents=intents)
-    await bot.add_cog(cog.WelcomeAndCoC(bot))
+    await bot.add_cog(utility_cog.Utility(bot))
+    await bot.add_cog(welcome_and_coc_cog.WelcomeAndCoC(bot))
     await bot.start(config.DISCORD_TOKEN)
