+++ bot/config.py
@@ -34,3 +34,4 @@ DISCORD_GUILD = get_env_var("DISCORD_GUILD")
 MEMBER_ROLE_NAME = get_env_var("MEMBER_ROLE_NAME", "members")
 COC_MESSAGE_ID = get_env_var_int("COC_MESSAGE_ID")
 COC_MESSAGE_LINK = get_env_var("COC_MESSAGE_LINK")
+WELCOME_CATEGORY_AND_CHANNEL_PREFIX = "welcome"
