+++ bot/messages.py
@@ -1,18 +1,27 @@
 NEW_MEMBER_MESSAGE = (
-    "Γεια σου @{name}, καλωσόρισες στο {guild}! 😊\n\n"
+    "Γειά σου {name}, καλωσόρισες στο {guild}! 😊\n\n"
     "Οταν βρεις λίγο χρόνο, σε παρακαλώ διάβασε τον [Κώδικα Δεοντολογίας μας]({link}) "
     "και αντίδρασε με thumbs-up (👍) στο μήνυμα. Μόλις το κάνεις, "
     "θα σου δώσω τον απαραίτητο ρόλο ώστε να δεις όλα τα υπόλοιπα κανάλια και να "
     "συμμετέχεις στη συζήτηση!\n\n"
     "---\n\n"
-    "Hey @{name}, thanks for joining {guild}! 😊\n\n"
+    "Hey {name}, thanks for joining {guild}! 😊\n\n"
     "Whenever you find some time, please go through [our Code of Conduct]({link}) "
     "and react with a thumbs-up (👍) on the message. As soon as you do, I'll give you "
     "permissions to see all the other channels and join in on the fun! "
 )
 
 ALREADY_EXISTS_MESSAGE = (
-    "Γεια σου @{name}, καλωσόρισες ξανά στο {guild}! Χαίρομαι που είσαι πάλι μαζί μας! 😊\n\n"
+    "Γειά σου {name}, καλωσόρισες ξανά στο {guild}! Χαίρομαι που είσαι πάλι μαζί μας! 😊\n\n"
+    "Μπορείς να αντιδράσεις με thumbs-up (👍) στον [Κώδικα Δεοντολογίας μας]({link}) "
+    "για να σου δώσω τον απαραίτητο ρόλο ώστε να δεις όλα τα υπόλοιπα κανάλια και να "
+    "συμμετέχεις στη συζήτηση!\n\n"
+    "---\n\n"
+    "Hey {name}, we see that you re-joined {guild}! It's great to have you back! 😊\n\n"
+    "You can react with a thumbs-up (👍) to [our Code of Conduct]({link}). "
+    "As soon as you do, I'll give you permissions to see all the other channels "
+    "and join in on the fun! "
+)
     "---\n\n"
     "Hey @{name}, we see that you re-joined {guild}! It's great to have you back! 😊"
 )
